refactor(phase-3b): extract JiraWorkflowService
Continues the jira_client.py god-class decomposition (ADR-002).
Four workflow-related methods (3 public + 1 private) move from
``jira_client.py`` into a focused ``JiraWorkflowService`` exposed
as ``self.workflows``. ``JiraClient`` keeps a thin delegator for
each public method; the private helper is fully internal to the
service.

Methods moved
-------------
- get_workflow_schemes (public)
- _get_workflow_schemes_per_project (private fallback used by above)
- get_workflow_transitions (public)
- get_workflow_statuses (public)

Numbers
-------
- jira_client.py: 2610 -> 2467 LOC
- jira_workflow_service.py: 0 -> 204 LOC

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean
- ruff check / ruff format: clean
